backend/requirements.txt
asgiref==3.5.0
Django==3.2.13
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
Pillow==9.1.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
